backend/requirements.txt
Django==5.2.6
Pillow==11.3.0
behave==1.2.6
colorama==0.4.6
iniconfig==2.0.0
packaging==24.2
parse==1.20.2
parse_type==0.6.4
pluggy==1.5.0
pytest==8.3.4
six==1.17.0
ReportLab==4.2.2
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.4.0
weasyprint==61.2
